@@ -5,7 +5,7 @@ package base
 // reference model.
 // https://specifications.openehr.org/releases/RM/latest/common.html#_archetyped_class
 type Archetyped struct {
-	Type        string    `json:"_type"`
+	Type        ItemType  `json:"_type"`
 	ArchetypeID ObjectID  `json:"archetype_id"`
 	TemplateID  *ObjectID `json:"template_id,omitempty"`
 	RmVersion   string    `json:"rm_version"`
